+++ include/framewatch/hooks/present_hook.h
@@ -11,6 +11,7 @@ namespace framewatch {
 enum class HookBackend {
     None,
     Dx11,
+    Vulkan,
 };
 
 enum class HookState {